Move from single execute flag to stack
The way this works is it stores a stack of true/false, and depending on
what it currently is, runs or doesn't run. It massively simplifies
branching and makes it possible to go infinitely. I will probably make
something demonstrating a few layers of possible branching later, but
I'm tired and this was a quick hack that I'm surprised worked.